macos: merge AppController and AppDelegate, organize groups
