--- macos/Sources/main.swift
@@ -1,7 +1,0 @@
-import AppKit
-
-let app = NSApplication.shared
-let delegate = AppDelegate()
-app.delegate = delegate
-
-_ = NSApplicationMain(CommandLine.argc, CommandLine.unsafeArgv)
